api/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy[asyncio]>=2.0
asyncpg>=0.30.0
alembic>=1.14.0
pydantic>=2.0
pydantic-settings>=2.0
python-multipart>=0.0.9
bcrypt>=4.0.0
paho-mqtt>=2.0.0
